% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbEst.R
\name{getAllMBs}
\alias{getAllMBs}
\title{Estimate Markov Blankets of a vector of target nodes}
\usage{
getAllMBs(
  targets,
  dataset,
  threshold = 0.01,
  method = "MMPC",
  test = "testIndFisher",
  verbose = TRUE
)
}
\arguments{
\item{targets}{is a vector of integers which will identify the target nodes of our algorithm}

\item{dataset}{A dataset used for estimation}

\item{threshold}{A positive real number designating the threshold for conditional independence tests used in the algorithm}

\item{method}{The name of the Markov Blanket estimation algorithm to use. Valid algorithms include "MMPC", "SES", and "gOMP".}

\item{test}{The conditional independence test to use in the algorithm. Default is the Fisher Independence test.}

\item{verbose}{Whether to provide detailed output}
}
\description{
`getAllMBs()` applies Markov Blanket algorithms to a vector of target nodes and
their first-order neighbors
}
