% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbEst.R
\name{getMB}
\alias{getMB}
\title{Estimate Markov Blanket of Target Node}
\usage{
getMB(
  target,
  dataset,
  threshold = 0.01,
  method = "MMPC",
  test = "testIndFisher",
  verbose = FALSE
)
}
\arguments{
\item{target}{An integer identifying the target node whose MB we are attempting to recover}

\item{dataset}{A dataset used for estimation}

\item{threshold}{A positive real number designating the threshold for conditional independence tests used in the algorithm}

\item{method}{The name of the Markov Blanket estimation algorithm to use. Valid algorithms include "MMPC", "SES", and "gOMP".}

\item{test}{The conditional independence test to use in the algorithm. Default is the Fisher Independence test.}

\item{verbose}{Whether to provide detailed output}
}
\description{
`getMB()` applies a Markov Blanket estimation algorithm to a target node
}
