\name{LocalFCI-package}
\alias{LocalFCI-package}
\alias{LocalFCI}
\docType{package}
\title{\packageTitle{LocalFCI}}
\description{\packageDescription{LocalFCI}}
\details{
\tabular{ll}{
Package: \tab LocalFCI \cr
Type: \tab Package \cr
Version: \tab 1.0 \cr
Date: \tab 2021-10-20 \cr
License: \tab None \cr
}
The primary function to use is localfci. The user supplies a set of target nodes, and the function will return an estimate of the structure around those nodes.
}
\section{Package Content}{\packageIndices{LocalFCI}}
\author{\packageAuthor{LocalFCI}}
\section{Maintainer}{\packageMaintainer{LocalFCI}}
\references{

}
\keyword{package}
\seealso{

}
\examples{

}
