% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localpc.R
\name{localpc}
\alias{localpc}
\title{Local PC Algorithm}
\usage{
localpc(
  data = NULL,
  true_dag = NULL,
  targets,
  node_names = NULL,
  lmax = 3,
  tol = 0.01,
  mb_tol = 0.05,
  method = "MMPC",
  test = "testIndFisher",
  verbose = TRUE
)
}
\arguments{
\item{data}{A dataframe or matrix containing network data}

\item{true_dag}{An optional parameter used primarily for simulations 
in order to obtain the Markov Blankets of certain target nodes}

\item{targets}{A numeric vector specifying the target nodes}

\item{node_names}{A string vector with the node names, useful for tracking output}

\item{lmax}{The maximum size of separating set considered in the algorithm. Default is 3}

\item{tol}{The significance level of our conditional independence tests. Default is 0.01}

\item{mb_tol}{The significance level of our Markov Blanket estimation algorithms. Default is 0.05}

\item{method}{Algorithm used for Markov Blanket estimation. Default is MMPC.}

\item{test}{The conditional independence test used for all algorithms.}

\item{verbose}{A boolean which determines whether or not to print a trace of the output.}
}
\description{
`localpc()` applies the Local PC algorithm to a vector of target nodes in a
network structure and returns an estimated PDAG
}
