\name{LocalFCI-package}
\alias{LocalFCI-package}
\alias{LocalFCI}
\docType{package}
\title{\packageTitle{LocalFCI}}
\description{\packageDescription{LocalFCI}}
\section{Package Content}{\packageIndices{LocalFCI}}
\author{\packageAuthor{LocalFCI}}
\section{Maintainer}{\packageMaintainer{LocalFCI}}
\keyword{package}
